more useful scripts!

+++ scripts/setup_kurome_dependencies.sh
@@ -11,20 +11,21 @@ fi
 # update git submodules
 bash $BENDER_SCRIPTS_DIR/update_submodules.sh
 
-cd $BENDER_DEPENDENCIES_DIR
 
 # download ceres
-rm -rf ceres-solver-2.2.0
-wget ceres-solver.org/ceres-solver-2.2.0.tar.gz
-gunzip ceres-solver-2.2.0.tar.gz
-tar xf ceres-solver-2.2.0.tar
-rm -rf ceres-solver-2.2.0.tar
+cd $BENDER_DEPENDENCIES_DIR
+if [ ! -d ceres-solver-2.2.0 ]; then
+   wget ceres-solver.org/ceres-solver-2.2.0.tar.gz
+   gunzip ceres-solver-2.2.0.tar.gz
+   tar xf ceres-solver-2.2.0.tar
+   rm -rf ceres-solver-2.2.0.tar
+fi
 
 # install all the other stuff we can grab from the web
-sudo apt install libabsl-dev
-sudo apt install libgoogle-glog-dev
-sudo apt install libsfml-dev
-sudo apt install libeigen3-dev
+sudo apt install libabsl-dev -y
+sudo apt install libgoogle-glog-dev -y
+sudo apt install libsfml-dev -y
+sudo apt install libeigen3-dev -y
 
 # build and install ceres
 cd $BENDER_DEPENDENCIES_DIR/ceres-solver-2.2.0
